Move language identification to a dedicated file
